telegram_hosted_bot/app/requirements.txt
python-telegram-bot
requests
pyjwt[crypto]
boto3
requests-oauthlib
openai
python-dotenv
aiohttp